Move some files for correct code layering
