+++ src/sql/workbench/contrib/welcome/gettingStarted/electron-browser/gettingStarted.ts
