--- src/sql/workbench/update/electron-browser/gettingStarted.contribution.ts
@@ -1,9 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the Source EULA. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { ShowGettingStartedAction } from 'sql/workbench/update/electron-browser/gettingStarted';
-import { registerAction2 } from 'vs/platform/actions/common/actions';
-
-registerAction2(ShowGettingStartedAction);
